# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "UGC Studio API",
  "description": "Local AI video production: projects, renders, voice, timeline, fixes, exports.",
  "version": "1.0"
 },
 "paths": {
...
  "/api/uploads": {
   "post": {
    "summary": "Upload",
    "operationId": "upload_api_uploads_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_api_uploads_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Upload Api Uploads Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/personas": {
   "get": {
    "summary": "Personas List",
    "operationId": "personas_list_api_personas_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "additionalProperties": true,
          "type": "object"
         },
         "type": "array",
         "title": "Response Personas List Api Personas Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "summary": "Personas Create",
    "operationId": "personas_create_api_personas_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PersonaIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Personas Create Api Personas Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/personas/{name}/files/{file}": {
   "get": {
    "summary": "Persona File",
    "operationId": "persona_file_api_personas__name__files__file__get",
    "parameters": [
     {
      "name": "name",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Name"
      }
     },
     {
      "name": "file",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "File"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/site": {
   "post": {
    "summary": "Site Analyze",
    "operationId": "site_analyze_api_site_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SiteIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Site Analyze Api Site Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects": {
   "get": {
    "summary": "Projects",
    "operationId": "projects_api_projects_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "additionalProperties": true,
          "type": "object"
         },
         "type": "array",
         "title": "Response Projects Api Projects Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "summary": "Project Create",
    "operationId": "project_create_api_projects_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreateIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Project Create Api Projects Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/import": {
   "post": {
    "summary": "Project Import",
    "description": "Create a project directly from a project definition (e.g. a template or an edited copy).",
    "operationId": "project_import_api_projects_import_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ImportIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Project Import Api Projects Import Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}": {
   "get": {
    "summary": "Project Get",
    "operationId": "project_get_api_projects__pid__get",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Project Get Api Projects  Pid  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "summary": "Project Put",
    "operationId": "project_put_api_projects__pid__put",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true,
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Project Put Api Projects  Pid  Put"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "summary": "Project Delete",
    "description": "Moves the project to outputs/_trash (recoverable), never deletes files.",
    "operationId": "project_delete_api_projects__pid__delete",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Project Delete Api Projects  Pid  Delete"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/plan": {
   "get": {
    "summary": "Project Plan",
    "operationId": "project_plan_api_projects__pid__plan_get",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Project Plan Api Projects  Pid  Plan Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/status": {
   "get": {
    "summary": "Project Status",
    "operationId": "project_status_api_projects__pid__status_get",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Project Status Api Projects  Pid  Status Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/media": {
   "get": {
    "summary": "Project Media",
    "description": "URLs of what the UI shows: final videos, cover, contact sheet, per-scene thumbnails and keyframes.",
    "operationId": "project_media_api_projects__pid__media_get",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Project Media Api Projects  Pid  Media Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/files/{path}": {
   "get": {
    "summary": "Project File",
    "operationId": "project_file_api_projects__pid__files__path__get",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     },
     {
      "name": "path",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Path"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{pid}/providers": {
   "get": {
    "summary": "Project Providers",
    "operationId": "project_providers_api_projects__pid__providers_get",
    "parameters": [
     {
      "name": "pid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Pid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response Project Providers Api Projects  Pid  Providers Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_upload_api_uploads_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_upload_api_uploads_post"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "ImportIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 60,
      "minLength": 1,
      "title": "Name"
     },
     "project": {
      "additionalProperties": true,
      "type": "object",
      "title": "Project"
     }
    },
    "type": "object",
    "required": [
     "name",
     "project"
    ],
    "title": "ImportIn"
   },
...
   "PersonaIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 40,
      "minLength": 1,
      "title": "Name"
     },
     "description": {
      "type": "string",
      "minLength": 5,
      "title": "Description"
     },
     "uploads": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Uploads",
      "default": []
     },
     "voice_style": {
      "type": "string",
      "title": "Voice Style",
      "default": ""
     },
     "language": {
      "type": "string",
      "title": "Language",
      "default": "English"
     }
    },
    "type": "object",
    "required": [
     "name",
     "description"
    ],
    "title": "PersonaIn"
   },
...
   "SiteIn": {
    "properties": {
     "url": {
      "type": "string",
      "pattern": "^https?://",
      "title": "Url"
     }
    },
    "type": "object",
    "required": [
     "url"
    ],
    "title": "SiteIn"
   },
...
  }
 }
}